[WAUM][Woo][1/n] Add scaffold tab for whatsapp utility messaging
### Changes proposed in this Pull Request:

Add a new scaffolding whatsapp utility tab to facebook for woocommerce plugin.

Note: The content and design are not finalized, these changes will be done in follow up diffs

### Screenshots:

<img width="1920" alt="Screenshot 2025-02-26 at 1 13 40 PM" src="https://github.com/user-attachments/assets/a03a9d4f-108a-41fd-ba90-429b2ff11e10" />

### Detailed test instructions:
1. Open Marketing > Facebook tab in wordpress containing facebook-for-woocommerce screens
2. Check if new whatsapp utility tab is visible
3. Clicking on the tab shows connect to whatsapp button

### Additional details:
Note: The feature flag is currently a boolean, we will implement a flagging mechanism which will be updated in following diffs.
Figma link:
https://www.figma.com/design/y6QWpbs7Ce8QcU1PjePpbw/3P-Integrations?node-id=567-97183&m=dev