Release/3.4.2 (#2946)
* Update README.md - Added noification for ownership transfer (#2910)

Summary:
This is a update to notify developers that the facebook-for-woocommerce plugin is now part of Meta and development is taken care by the engineers at Meta. We also have encouraged for contributions on this repo.

### Changes proposed in this Pull Request:

This PR contains mostly the changes around **README.md**.
Since mostly developers of any kind prefer looking at the **README** of a GitHub repo.
We have now changed the README emphasising the changes of the new **Owneship of the plugin** by Meta.
And the invitation for **opensource community** for their contributions.

### Screenshots:

 {F1975511490}

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/2910

Reviewed By: mshymon, vinkmeta

Differential Revision: D70323913

Pulled By: SayanPandey

fbshipit-source-id: 3fd9937d20f968132dba4fcc23ebfcabae9a3f55

* Changing APP to PLUGIN - README.MD (#2916)

Summary:
Changing the app word to plugin to more suit with the idea of facebook-for-woocommerce

### Changes proposed in this Pull Request:

This is a simple change of APP to PLUGIN in README.md text .

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/2916

Reviewed By: crisojog, vinkmeta

Differential Revision: D70486710

Pulled By: SayanPandey

fbshipit-source-id: 47eb9d84a32712d678e153d5b1a98fbf7074af98

* Removed the redundant entries from readme.txt and changelog.txt

* Updating version to 3.4.2

---------

Co-authored-by: Sayan Pandey <sayanpandey@meta.com>
Co-authored-by: Vahid Kaykhaei <vahidkay@meta.com>