Relaxing sync validation (#2969)
Summary:
## Description

Please include a summary of the changes and the related issue.
Please also include relevant motivation and context.
List any dependencies that are required for this change.

### Type of change

Please delete options that are not relevant.

- [X] New feature (non-breaking change which adds functionality)

| # | Field | Validation | Change |
| --- | --- | --- | ---|
| 1 | Description | ALL Cap | Remove |
| 2 | Description | length(5000) | Remove |
| 3 | Price | If missing | Remove|
| 4 | Title | length(150) | Remove |
| 5 | validate_product_visibility | if hidden| Keep for now (we need to sync items as hidden in case they are not visible) |
| 6 | Out of stock | if woocommerce_hide_out_of_stock_items we don't sync | Removed in another PR |
| 7 | product_status | ALL Cap | Keep (part of other efforts) |
| 8 | validate_product_sync_field| | Keep |
| 9 | validate_product_terms| Check whether the product's categories or tags (terms) exclude it from sync.| Keep |

Note: I am not super sure about (5) validate_product_visibility, I have removed it, but the item is being synced as active

## Screenshots
Please provide screenshots or snapshots of the system/state both before and after implementing the changes, if appropriate
### Before
In the before nothing synced showing error in the meta_box component

### After
<img width="1280" alt="image" src="https://github.com/user-attachments/assets/195d2190-cbbb-493c-aee4-94a9fb0bbaf8" />
https://www.facebook.com/commerce/catalogs/955650953442091/products

As you can see, I have synced items with different validations to ensure syncing is happening

## Test instructions

Create a product with a validation issue and try saving and check if it's being synced or not

## Checklist

- [X] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc))
- [X] I have commented my code, particularly in hard-to-understand areas
- [X] I have added tests and all the new and existing unit tests pass locally with my changes
- [X] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.

## Changelog entry

Relaxing sync validations

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/2969

Reviewed By: vinkmeta

Differential Revision: D71912613

Pulled By: raymon1

fbshipit-source-id: 8e358749664d426f809711a889a6665d257f9a4c