Remove UI of a checkbox that controls enablement of the new style feed generation. (#3056)
Summary:
## Description

As of now there is a switch in UI that controls enablement of the new style feed generation.

The plan is to remove this from the UI, while we will still respect the value that was set.
We are planning for the new style of feed generation to be additionally tested and then rolled out via a Gatekeeper to insure consistent feed generation across all sellers.

We are removing this from the UI since this is exposing an internal infra change and there will be no need for seller to be aware or have a control over it.
The control was removed from both old and new plugin page UX.

### Type of change

Please delete options that are not relevant.

- New feature (non-breaking change which adds functionality)

## Checklist

- [] I have commented my code, particularly in hard-to-understand areas.
- [] I have confirmed that my changes do not introduce any new PHPCS warnings or errors.
- [] I have checked plugin debug logs that my changes do not introduce any new PHP warnings or FATAL errors.
- [] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc)).
- [] I have added tests (if necessary) and all the new and existing unit tests pass locally with my changes.
- [] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.
- [] I have updated or requested update to plugin documentations (if necessary). Meta employees to follow this [wiki]([url](https://fburl.com/wiki/nhx73tgs)).

## Changelog entry

One liner entry to be surfaced in changelog.txt

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3056

Test Plan:
Was manually tested, screenshots provided below.

## Screenshots
Please provide screenshots or snapshots of the system/state both before and after implementing the changes, if appropriate
### Before

<img width="1277" alt="Screenshot 2025-04-14 at 15 37 02" src="https://github.com/user-attachments/assets/6f00e56c-1313-432f-9a2e-41adef32a39f" />

### After

The UI control has been removed:
<img width="1481" alt="Screenshot 2025-04-14 at 15 32 03" src="https://github.com/user-attachments/assets/284c138a-dce8-4c8a-95ae-c9b257c1601b" />

Reviewed By: ajello-meta, vinkmeta

Differential Revision: D72970938

Pulled By: mshymon

fbshipit-source-id: 3ada315d22168a0cd0c8ff3723351898052a97e0

(cherry picked from commit 9596404b8409baca3653a07265f6dc418c651180)