Release plugin workflow (#3309)
Summary:
This workflow deploys the plugin to Wordpress SVN and creates a new GitHub release.
It is using the Deployment environment and will only deploy when pushing to release/current branch which is a protected branch.
It uses the package.json version to get the latest version.
On SVN it replaces trunk with the contents of the built package and creates a new tag based on the version.
It also creates a new GitHub prerelease using the readme.txt changelog.

It currently validates both git and svn tags before moving forward with the release process.

**It does not set the "Stable tag" on readme.txt**

Succesful test run:
https://github.com/facebook/facebook-for-woocommerce/actions/runs/15544252114

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3309

Reviewed By: vahidkay-meta

Differential Revision: D76333813

Pulled By: tzahgr

fbshipit-source-id: 896ebb1c84a3ef5361dd06f763de0a286e7fb118