fix: rollout-switch-transient-flag (#3595)
Summary:
## Description

T227891605 T227892775 T237040935

What are we solving?
As quoted in the task

"The issue seems to be because of how the RolloutSwitch is setup. The RolloutSwitches are fetched every 60 min and stored in wp_options table, a user who has installed an old version of the plugin will have all the old active switches fetched and stored in table for 60 min.
If the plugin upgrade happens before the transient expires, then the new switches will be in the active_switches const but will not be in the wp_options DB which is out of date since we do not fetch the rolloutswitches through api call.
This results in returning true here since an active switch is not available in the wp_options table.
"

This change ensures RolloutSwitch Class **doesn't return early with stale transient flag.**

**added a suffix of "plugin version" to  _wc_facebook_for_woocommerce_rollout_switch_flag_**  transient flag to allow Rollout Switches to be updated with a new version release.

Fetched the plugin version using

`plugin_version = $this->plugin->get_version();`

which gets the value `const PLUGIN_VERSION` defined in facebook-for-woocommerce.php file.

### Type of change
- Fix (non-breaking change which fixes an issue)

## Checklist

- [yes] I have commented my code, particularly in hard-to-understand areas, if any.
- [yes] I have confirmed that my changes do not introduce any new PHPCS warnings or errors.
- [yes] I have checked plugin debug logs that my changes do not introduce any new PHP warnings or FATAL errors.
- [yes] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc)).
- [] I have added tests (if necessary) and all the new and existing unit tests pass locally with my changes.
- [] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.
- [] I have updated or requested update to plugin documentations (if necessary). Meta employees to follow this [wiki]([url](https://fburl.com/wiki/nhx73tgs)).

## Changelog entry

added a suff _wc_facebook_for_woocommerce_rollout_switch_flag_  transient flag to allow Rollout Switches to be updated with a new version release.

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3595

Test Plan:
Please describe the tests that you ran to verify your changes.
Provide instructions so we can reproduce.
Please also list any relevant details for your test configuration.

## Screenshots
Transitent values are set in wp_options table, screenshots from the table to show updated wp_options with new plugin version
### Before {F1981842031, width=300}

### After
with 3.5.7 version
 {F1981842059, width=300}

when version updated, it doesn't return at
		`if ( 'yes' === get_transient( $flag_name ) ) {`
			`return;`
		}

Instead, sets a new transient value and continue on
 {F1981842121, width=300}

Reviewed By: vinkmeta, ajello-meta

Differential Revision: D82036669

Pulled By: rithikb24

fbshipit-source-id: 03bc4e4e40b43aa076a95f938c0bc352c7e23d27