Adding Profit Margin signal to AddToCart and Purchase events (#3730)
Summary:
## Description

Added support for calculation of Profit Margin to the Purchase event. The way it works is if you have COGS enabled on WooCommerce ( WooCommerce settings > Advanced > Features ) or you have WPFactory Cost-of-Goods plugin installed and have COGS value added to a Simple or Variable product. In such case, the plugin picks the cogs values up and calculates the profit margin for the order, and sends it through net_revenue field of the Pixel&CAPI events. We prioritize WooC Cogs over WPFactory. So if the value is set for both WooC COGS and WPFactory, the plugin will pick up the WooC value.

The Profit margin is calculated only if all of the products in an order have COGS value set for them.

### Type of change

Please delete options that are not relevant

Add (non-breaking change which adds functionality)

## Checklist

- [x] I have commented my code, particularly in hard-to-understand areas, if any.
- [x] I have confirmed that my changes do not introduce any new PHPCS warnings or errors.
- [x] I have checked plugin debug logs that my changes do not introduce any new PHP warnings or FATAL errors.
- [x] I followed general Pull Request best practices. Meta employees to follow this [wiki]([url](https://fburl.com/wiki/2cgfduwc)).
- [x] I have added tests (if necessary) and all the new and existing unit tests pass locally with my changes.
- [x] I have completed dogfooding and QA testing, or I have conducted thorough due diligence to ensure that it does not break existing functionality.
- [x] I have updated or requested update to plugin documentations (if necessary). Meta employees to follow this [wiki]([url](https://fburl.com/wiki/nhx73tgs)).

## Changelog entry

Added support for Profit Margin to the Purchase event.

Pull Request resolved: https://github.com/facebook/facebook-for-woocommerce/pull/3730

Test Plan:
Various Unit and Integration tests have been added to ensure the functionality works as expected. The integration tests enable/disable both of the WooC & WPFactory COGS functionalities and evaluate that our CostOfGoods class calculates the values correctly for a product or a set of products.
These tests are added to:
tests/Unit/Integrations/CostOfGoods
tests/integration/COGS

The integration tests files start with numbers. This is a hack to ensure tests run in order so that the side-effects are minimized.

The functionality has also been tested manually with only WooC, only WPFactory, and both integrations being present. I tested Simple and Variable products, in situations where COGS value existed or was missing.

Reviewed By: rafael-curran

Differential Revision: D88930239

Pulled By: vahidkay-meta

fbshipit-source-id: 45720db3e4b61e1f41d8c16cde54bf487e6eabbe